pyproject.toml
[project]
name = "ta3lem"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-braces>=1.17.0",
    "django>=5.2.6",
    "django-embed-video>=1.4.10",
    "pymemcache>=4.0.0",
    "django-debug-toolbar>=6.0.0",
    "redis>=6.4.0",
    "django-redisboard>=9.0.0",
    "pillow>=11.3.0",
    "honcho>=2.0.0",
    "django-htmx>=1.26.0",
    "psycopg2-binary>=2.9.11",
    "nh3>=0.3.2",
    "markdown>=3.10",
    "py-gfm>=2.0.0",
    "mdx-cite",
    "markdown-emdash>=0.2.0",
    "mdx-unimoji>=1.1",
    "djlint>=1.36.4",
    "gunicorn>=23.0.0",
    "uvicorn>=0.40.0",
    "python-dotenv>=1.2.1",
    "whitenoise[brotli]>=6.11.0",
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.28.0",
    "djangorestframework-simplejwt>=5.4.0",
    "django-filter>=25.1",
]

[tool.uv.sources]
mdx-cite = { git = "https://github.com/triopter/mdx-3-cite.git" }
